refactor: E2E Test 중복 코드 제거
